Squashed commit of the following:
commit e6b741e91c1a0c76cd680f488bb1e2c165710632
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Tue Nov 19 21:07:30 2024 +0200

    Mv NewLectureFrom to LectureFrom with tweaks
